refactor: rewrite code with using classes
