+++ index.html
@@ -7,6 +7,6 @@
 	<link rel="stylesheet" href="src/style.css">
 </head>
 <body>
-	<script src="src/index.js"></script>
+	<script src="src/index.js" type="module"></script>
 </body>
 </html>
