+++ src/js/data.js
@@ -1,0 +1,68 @@
+const DATA = {
+  Backquote: ['button', '`', '~', 'ё', 'Ё'],
+  Digit1: ['button', '1', '!', '1', '!'],
+  Digit2: ['button', '2', '@', '2', '"'],
+  Digit3: ['button', '3', '#', '3', '№'],
+  Digit4: ['button', '4', '$', '4', ';'],
+  Digit5: ['button', '5', '%', '5', '%'],
+  Digit6: ['button', '6', '^', '6', ':'],
+  Digit7: ['button', '7', '&', '7', '?'],
+  Digit8: ['button', '8', '*', '8', '*'],
+  Digit9: ['button', '9', '(', '9', '('],
+  Digit0: ['button', '0', ')', '0', ')'],
+  Minus: ['button', '-', '_', '-', '_'],
+  Equal: ['button', '=', '+', '=', '+'],
+  Backspace: ['long_button', '', '', '', ''],
+  Tab: ['button', '\t', '\t', '\t', '\t'],
+  KeyQ: ['button', 'q', 'Q', 'й', 'Й'],
+  KeyW: ['button', 'w', 'W', 'ц', 'Ц'],
+  KeyE: ['button', 'e', 'E', 'у', 'У'],
+  KeyR: ['button', 'r', 'R', 'к', 'К'],
+  KeyT: ['button', 't', 'T', 'е', 'Е'],
+  KeyY: ['button', 'y', 'Y', 'н', 'Н'],
+  KeyU: ['button', 'u', 'U', 'г', 'Г'],
+  KeyI: ['button', 'i', 'I', 'ш', 'Ш'],
+  KeyO: ['button', 'o', 'O', 'щ', 'Щ'],
+  KeyP: ['button', 'p', 'P', 'з', 'З'],
+  BracketLeft: ['button', '[', '{', 'х', 'Х'],
+  BracketRight: ['button', ']', '}', 'ъ', 'Ъ'],
+  Backslash: ['button', '\\', '|', '\\', '/'],
+  Delete: ['button', '', '', '', ''],
+  CapsLock: ['long_button', '', '', '', ''],
+  KeyA: ['button', 'a', 'A', 'ф', 'Ф'],
+  KeyS: ['button', 's', 'S', 'ы', 'Ы'],
+  KeyD: ['button', 'd', 'D', 'в', 'В'],
+  KeyF: ['button', 'f', 'F', 'а', 'А'],
+  KeyG: ['button', 'g', 'G', 'п', 'П'],
+  KeyH: ['button', 'h', 'H', 'р', 'Р'],
+  KeyJ: ['button', 'j', 'J', 'о', 'О'],
+  KeyK: ['button', 'k', 'K', 'л', 'Л'],
+  KeyL: ['button', 'l', 'L', 'д', 'Д'],
+  Semicolon: ['button', ';', ':', 'ж', 'Ж'],
+  Quote: ['button', '\'', '"', 'э', 'Э'],
+  Enter: ['long_button', '\n', '\n', '\n', '\n'],
+  ShiftLeft: ['long_button', '', '', '', ''],
+  KeyZ: ['button', 'z', 'Z', 'я', 'Я'],
+  KeyX: ['button', 'x', 'X', 'ч', 'Ч'],
+  KeyC: ['button', 'c', 'C', 'с', 'С'],
+  KeyV: ['button', 'v', 'V', 'м', 'М'],
+  KeyB: ['button', 'b', 'B', 'и', 'И'],
+  KeyN: ['button', 'n', 'N', 'т', 'Т'],
+  KeyM: ['button', 'm', 'M', 'ь', 'Ь'],
+  Comma: ['button', ',', '<', 'б', 'Б'],
+  Period: ['button', '.', '>', 'ю', 'Ю'],
+  Slash: ['button', '/', '?', '.', ','],
+  ArrowUp: ['button', '', '', '', ''],
+  ShiftRight: ['long_button', '', '', '', ''],
+  ControlLeft: ['button', '', '', '', ''],
+  AltLeft: ['button', '', '', '', ''],
+  MetaLeft: ['button', '', '', '', ''],
+  Space: ['space', ' ', ' ', ' ', ' '],
+  AltRight: ['button', '', '', '', ''],
+  ArrowLeft: ['button', '', '', '', ''],
+  ArrowDown: ['button', '', '', '', ''],
+  ArrowRight: ['button', '', '', '', ''],
+  ControlRight: ['button', '', '', '', ''],
+};
+
+export default DATA;
